### 2020-05-21 #2189 @whedon
[PRE REVIEW]: uravu: Making Bayesian modelling easy(er)
**Submitting author:** @arm61 (<a href="http://orcid.org/0000-0003-3381-5911">Andrew McCluskey</a>)
**Repository:** <a href="https://github.com/arm61/uravu" target ="_blank">https://github.com/arm61/uravu</a>
**Version:** v0.0.7
