### 2020-05-22 53d41ae
make model line smoother
