### 2020-05-23 #2214 @ejhigson
Hi @arm61 - this is a really neat piece of software, nice work!

I have a few minor questions about the installation process. Your documentation says:
```
pip install -r requirements.txt
python setup.py build
python setup.py install
pytest
```
Is there a reason that the separate ```python setup.py build``` step is needed? You could also consider putting the requirements into the ```install_requires``` property in ``setup.py`` to eliminate the first step too. It might be nice to list ``uravu`` on PyPI so it can be easily installed with pip, but this is optional.

I initially got a bunch of the following errors running tests
```
E           AttributeError: module 'scipy.misc' has no attribute 'logsumexp'
```
which turned out to be due to me having an older version of ```dynesty``` (0.9.2) which used an old ```scipy``` function which is no longer included in the latest version of ```scipy```. They were fixed when I upgraded ```dynesty``` (to 1.0.1). I think it would be useful to put a set of minimum versions in your requirements (e.g. dynesty>=1.0.1 or whatever the appropriate version is).
