### 2020-05-23 arm61/uravu#4 @arm61
Add dynasty greater than 1.0.1
https://github.com/openjournals/joss-reviews/issues/2214#issuecomment-633026018
