### 2020-05-23 #2214 @ejhigson
A minor fix in the documentation - I think the hyperlink to the contributing guidelines in README.md is broken?
