### 2020-05-23 #2214 @ejhigson
@arm61 please could you explain the contributions of the second author of the paper Tim Snow? I see that on the github repo you seem to be the only person to have contributed code(?)
