### 2020-05-23 arm61/uravu#6 @arm61
Mention why installation is different in docs
The docs installation guidance says `pip` install but the `README.md` says `python setup.py ...`. Add an explanation to the docs as to why these are different.
