### 2020-05-23 #2214 @arm61
Nice catch @ejhigson, I have now added `dynasty>=1.0.1` to the `requirements.txt` and to the `install_requires` and `setup_requires` of the `setup.py`, this is currently in [PR #5](https://github.com/arm61/uravu/pull/5) and which is now pulled.

The difference between the readthedocs and `README.md` on Github is that the former is meant to be the general purpose installation while the later is for those interested in a "development" build (I [will mention this](https://github.com/arm61/uravu/issues/6) in the installation guidance for improved clarity). This is also the reason why `pytest`, `matplotlib` and `seaborn` are in the `requirements.txt` and not the `setup.py` (as these modules are only **required** for testing of the [`plotting.py` module](https://github.com/arm61/uravu/blob/421d04ad6a3ee04634adb3bbfd67f081a668c493/uravu/plotting.py#L18) and not really needed for just using the code).

The logic of the `python setup.py build` is just that when I am doing development work I like to [segment things as much as possible](https://docs.python.org/2/install/#splitting-the-job-up).

I have made [an issue](https://github.com/arm61/uravu/issues/7) for the broken hyperlink in the README.md so will get that sorted ASAP.

Finally, Tim Snow is my supervisor at Diamond Light Source, while he did not contribute to the code base itself, his contributions in terms of discussion when developing the code and writing the paper were invaluable.
