### 2020-05-23 4813a8a
fixing broken hyperlink to contributing guidelines
