### 2020-05-23 5a69f55
add a clarification to the installation guidence
