### 2020-05-23 1df7df4
more clarity added to the README regarding installation also fix to formatting
