### 2020-05-23 #2214 @ejhigson
That's great, thank you very much @arm61! As I understand it Tim Snow's non-code contributions meet the authorship criteria (https://joss.readthedocs.io/en/latest/submitting.html#authorship) so that's all good - thank you for explaining.

Looking at the paper: I think there is a typo in the first sentence: "interface data analysis" -> "interface for data analysis"? Also do you know of any other similar pieces of software that act as data analysis interfaces like ``uravu``? I think it is worth mentioning any alternatives or explicitly stating that there are no other such packages to your knowledge.
