### 2020-05-23 #2214 @arm61
Thanks for catching that typo @ejhigson (it must have slipped through that particular gap in my brain!). I have resolved this and included a brief discussion of the uniqueness of `uravu` in [PR #11](https://github.com/arm61/uravu/pull/11).

As is clear from the addition I believe that `uravu` is the only package that offers Bayesian data modelling in such a simple interface. However, I am more than happy to be proved wrong if yourself, @joshspeagle, or @dfm know of any!
