### 2020-05-26 #2214 @ejhigson
Hi @arm61 - thank you very much for your responses!

I had a look over the code and API documentation and most things seemed to be nicely documented, but there are some functions where some of the arguments are missing documentation e.g. ``uravu.sampling.nested_sampling``. Please could you check over this and filling in the gaps? I also just saw a very minor typo in the paper where you are missing a space "matplotlib(Hunter".

I ran all of the getting started and tutorial jupyter notebooks which are nicely written - these all ran perfectly for me.

One other suggestion - I think it may be worth mentioning _dynamic_ nested sampling in the documentation and paper, since ``dynesty`` is nominally a dynamic nested sampling software package and I see you use dynamic nested sampling via the ``dynesty.DynamicNestedSampling`` sampler in ``uravu.sampling.nested_sampling``.
