### 2020-05-27 0c143da
improvements to docs strings
