### 2020-05-27 b4b2ea5
inclusion of references to dynamic nested sampling
