+++ paper/paper.bib
@@ -19,6 +19,27 @@
     doi = {10.1038/s41592-019-0686-2},
 }
 
+@ARTICLE{higson_dynamic_2019,
+    doi={10.1007/s11222-018-9844-0},
+    volume = {29},
+    pages = {891},
+    journal = {Stat. Comput.},
+    author = {Higson, E. and Handley, W. and Hobson, M. and Lasenby, A.},
+    title = {Dynamic nested sampling: an improved algorithm for parameter estimation and evidence calculation},
+    year={2019}
+}
+
+@ARTICLE{skilling_nested_2006,
+  doi = {10.1063/1.1835238},
+  volume = {735},
+  number = {1},
+  journal = {AIP Conference Proceedings},
+  pages = {395},
+  author = {Skilling, J.},
+  title = {{N}ested {S}ampling},
+  year = {2006}
+}
+
 @ARTICLE{salvatier_pymc3_2016,
     author = {Salvatier, John and Wiecki, Thomas V. and Fonnesbeck, Christopher},
     title = {Probabilistic programming in Python using PyMC3},
