### 2020-05-27 #2214 @arm61
Hey @ejhigson, thanks for the additional comments! The benefit of having someone independent look though the code is really necessary for finding things like missing argument docs!

I have improved the docstrings in https://github.com/arm61/uravu/commit/0c143daec8d749994abd8c9adfe5af5261776714. Hopefully I have got them all covered now!

The missing space is dealt with in https://github.com/arm61/uravu/commit/93ccdf5882dbd9b9c149a9f48bf089287cb0b091.

And finally, I totally agree that mentioning `uravu`'s ability to use `dynasty`'s dynamic nested sampling is important.
- I have added a mention in the paper (and some fresh references)
- I have mentioned dynamic nested sampling in the nested sampling tutorial
- and I have improved the docstrings for `Relationship.nested_sampling()` and `sampling.nested_sampling()` to mention dynamic nested sampling

These changes are all in https://github.com/arm61/uravu/commit/b4b2ea51550d6321f7e3df25875cb62f28724943 and https://github.com/arm61/uravu/commit/c8c8b0fd31646ed2e1d8371b20f6fc9ad6bd1e78

Please let me know if you catch anything else, and thanks for the kind words about the tutorial notebooks 😄.
