### 2020-05-27 #2214 @ejhigson
Hi @arm61 - thank you very much for these changes, they all look great to me!

I am satisfied that you have met all the reviewer criteria. Congratulations on a nice paper and an awesome software package!
