### 2020-05-27 567bda4
add dns to docs homepage
