+++ docs/source/index.rst
@@ -45,7 +45,7 @@ Getting and :py:class:`uravu.relationship.Relationship` running is as simple as.
 .. _Github: https://github.com/arm61/uravu
 .. _Goodman & Weare’s Affine Invariant Markov chain Monte Carlo (MCMC) Ensemble sampler: https://doi.org/10.2140/camcos.2010.5.65
 .. _nested sampling: https://doi.org/10.1063/1.1835238
-.. _nested sampling: https://doi.org/10.1007/s11222-018-9844-0
+.. _dynamic nested sampling: https://doi.org/10.1007/s11222-018-9844-0
 
 .. toctree::
    :hidden:
