### 2020-06-02 #2214 @joshspeagle
Hi Andrew,

Sorry for the delay in getting to this compared with @ejhigson. I've gone through the paper, code, tutorials, and docs and only have a few small comments on the docs, none of which are a barrier to publication (as indicated by the reviewer checklist above). These are:

- In the docs and tutorials, it would be nice to include a little more description of what's going on under the hood when you call various functions since the user interface is purposely designed to ease many of these subtleties. For instance, `straight_line` and `diff_evo` seem self-explanatory to me personally, but it would be good to highlight those and other options for users whose first interaction with the code (and possibly Bayesian statistics) will be running through the tutorials in `uravu`. You do this in some places in later tutorials but not as much earlier, so it'd be good to homogenize a bit.

- I really love how easy the code makes using user-defined functions!!! However, I don't see a particular tutorial on how those might need to be structured. I could imagine users having questions like: Do they require non-array arguments? Do they allow for additional arguments/keywords to be passed through `uravu`? Can they return an additional `blob`, or does it have to be just the output model? etc. etc. that would be easy to address in some way in the docs.

- Might be good to mention in the docs that `variable_medians` stands in for the modes when running `max_likelihood` but are actually the medians when using MCMC or NS. It might be worth adding a separate summary variable like `variable_map` since sometimes users might want both the median and the mode.

- During the NS tutorial, you have a comparison with the line "We can see that the highest order terms are quite small for the larger degree of polynomial." I don't actually see a direct comparison, however, but only the plot for the single-parameter fit. Did you mean to overplot the four models?

- It might be worth commenting that some of the outputs are piped through `emcee` or `dynesty`, and directing users there to get more information.

The most recent version of the paper is clear and meets all the standards for publishing in JOSS. The code is well-documented and easy to use. Great work on making an excellent package!
