### 2020-06-02 #2214 @arm61
Hi Josh!

Thanks for the great review and the detailed notes on the docs. Appreciated massively! I have opened them all as issues 😄

@dfm if you are happy, can we hold off until I get back to 0 issues (probably tomorrow) before moving forward? I would like to have the best possible `1.0.0` build to go along with the paper.
