### 2020-06-02 #2214 @dfm
@ejhigson, @joshspeagle: Also a big thanks from me for your reviews!!

@arm61: That sounds good. I also want to proofread the paper and do a bit of sanity checking before you tag your release for the publication so hold off on that for now, but let me know when you're satisfied with your responses to these suggestions.
