### 2020-06-03 d5d0eb0
add variable_modes function to relationship and ensure compatibility with the distribution dist_max property
