### 2020-06-03 c543d7f
improvements to the tutorials to be more clear in introductory
