### 2020-06-03 #2214 @arm61
Hey @dfm,

Thats me ticked off all of the points @joshspeagle rised. On his second bullet point, unfortunately `uravu` currently only supports **very** simple function definitions (no array args and only simplistic returns). However, I am happy with this functionality for a `1.0.0` build, and will work on adding this additional (quite complex) functionality in the future.

Please let me know when you are happy with everything and I will mint a `1.0.0` release.

Thanks again to @ejhigson and @joshspeagle for making the review process of JOSS a pleasure!
