### 2020-06-04 #2214 @dfm
@arm61: Everything is looking good to me! Can you now bump the version number and generate a new Zenodo release? **The title and author list of the Zenodo release must match the JOSS paper.** You can edit the metadata on the Zenodo page after making the release. Then please report the version number and Zenodo DOI back to this thread. Thanks!
