### 2020-06-04 #2214 @arm61
Version number: `1.0.0`

I will get the DOI as soon as possible [but zenodo.org is down just now](https://downforeveryoneorjustme.com/zenodo.org) 😱 .
