### 2020-06-04 #2214 @arm61
Too excited!

That's the old release deleted. Version number is still `1.0.0`, doi is now `10.5281/zenodo.3876888`
