### 2020-06-04 #2214 @kyleniemeyer
Hi @arm61, just a few quick edits on the paper:

- I think the semi-colon in the first sentence is not the right punctuation, instead a comma seems more appropriate
- in the first sentence after the bullets, "uravu also add the" should be "uravu also adds the"
- in the Statement of Need paragraph, "to the best of the author's knowledge" should either be "to the best of our knowledge" or "to the best of the authors' knowledge"
- in the emcee citation, please force capitalization of Python and MCMC by surrounding them with curly brackets, e.g., `{Python}`, `{MCMC}`
- same for Python and PyMC3 in the Salvatier 2016 citation

Please let me know when you've made those changes, and I'll get the article published!
