+++ paper/paper.bib
@@ -42,7 +42,7 @@
 
 @ARTICLE{salvatier_pymc3_2016,
     author = {Salvatier, John and Wiecki, Thomas V. and Fonnesbeck, Christopher},
-    title = {Probabilistic programming in Python using PyMC3},
+    title = {Probabilistic programming in {Python} using {PyMC3}},
     journal = {PeerJ Computer Science},
     volume = {2},
     pages = {e55},
@@ -65,8 +65,8 @@
         Jeremy and {Zuntz}, Joe and {Williams}, Peter and {Nelson}, Andrew 
         and {de Val-Borro}, Miguel and {Erhardt}, Tobias and {Pashchenko}, 
         Ilya and {Pla}, Oriol},
-    title = {emcee v3: A Python ensemble sampling toolkit for 
-        affine-invariant MCMC},
+    title = {emcee v3: A {Python} ensemble sampling toolkit for 
+        affine-invariant {MCMC}},
     journal = {Journal of Open Source Software},
     year = "2019",
     volume = {4},
