### 2020-06-05 #2214 @arm61
Hey @kyleniemeyer,

Thanks for the expert proof reading! Those changes are now in master (https://github.com/arm61/uravu/commit/14ad67ebb099c0e57b7d52ad1fdc7449a0b7b5ee).

Do I need to generate a new zenodo DOI with these changes?
