### 2020-06-05 #2214 @kyleniemeyer
@arm61 looks good! And no, if the software has not changed, the prior Zenodo archive is fine, since JOSS archives the paper itself.
